Moved Text contents to enum files and add fixes based on PR review.
