+++ tests/e2e/testData/translations.enum.ts
@@ -1,0 +1,5 @@
+/* eslint-disable @typescript-eslint/no-duplicate-enum-values */
+export enum EnglishTranslations {
+  ApproximateNumberDisclaimer = 'All numbers are approximate and meant to be used as guidance.',
+  ExposedPopulationInfoButtonDisclaimer = 'This layer shows the estimated rounded number of people potentially exposed per geographic area.',
+}
